Refactor CC index URL tutorial modules
